separate block and tx codec tests
